+++ next-docs/components/sidebar/RecursiveNavItem.tsx
@@ -1,4 +1,5 @@
 import { Accordion } from '@heathmont/moon-core-tw';
+import { ControlsChevronRight } from '@heathmont/moon-icons-tw';
 import Link from './Link';
 import type ItemType from './ItemType';
 
@@ -17,17 +18,27 @@ const RecursiveNavItem: React.FC<Props> = ({ item, pathname, onClick }) => {
   const isCurrent = href === gettingStartedPathname;
   if (children) {
     return (
-      <Accordion title={name} isContentInside={false}>
-        <div className="flex flex-col items-start gap-2 ps-8">
-          {children.map((subItem: ItemType) => (
-            <RecursiveNavItem
-              key={subItem.name}
-              item={subItem}
-              pathname={pathname}
-              onClick={onClick}
-            />
-          ))}
-        </div>
+      <Accordion>
+        <Accordion.Item value="item">
+          <Accordion.Header className="moon-open:[&_svg]:rotate-90">
+            <Accordion.Button>
+              <span>{name}</span>
+              <ControlsChevronRight className="text-trunks text-moon-16 transition-transform transition-200" />
+            </Accordion.Button>
+          </Accordion.Header>
+          <Accordion.ContentOutside>
+            <div className="flex flex-col items-start gap-2 ps-8">
+              {children.map((subItem: ItemType) => (
+                <RecursiveNavItem
+                  key={subItem.name}
+                  item={subItem}
+                  pathname={pathname}
+                  onClick={onClick}
+                />
+              ))}
+            </div>
+          </Accordion.ContentOutside>
+        </Accordion.Item>
       </Accordion>
     );
   }