+++ next-docs/public/examples/tagTW/LetterCases.tsx
@@ -1,0 +1,10 @@
+import { Tag } from '@heathmont/moon-core-tw';
+
+const Example = () => (
+  <>
+    <Tag>Default</Tag>
+    <Tag isUppercase={false}>Lowercase</Tag>
+  </>
+);
+
+export default Example;