+++ workspaces/core/src/tag/styles/TagContainer.tsx
@@ -3,9 +3,9 @@ import mergeClassnames from '../../utils/mergeClassnames';
 import setFontSize from '../private/utils/setFontSize';
 import setHeight from '../private/utils/setHeight';
 import setSpacing from '../private/utils/setSpacing';
-import type LabelProps from '../private/types/LabelProps';
+import type TagProps from '../private/types/TagProps';
 
-const LabelContainer: React.FC<LabelProps> = ({
+const TagContainer: React.FC<TagProps> = ({
   size,
   iconLeft,
   iconRight,
@@ -28,4 +28,4 @@ const LabelContainer: React.FC<LabelProps> = ({
   </div>
 );
 
-export default LabelContainer;
+export default TagContainer;