+++ packages/select/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @heathmont/moon-select
 
+## 9.16.0
+
+### Minor Changes
+
+- Accordeion, Tag, Tooltip
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@9.16.0
+  - @heathmont/moon-themes@9.16.0
+  - @heathmont/moon-utils@9.16.0
+
 ## 9.15.2
 
 ### Patch Changes